% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cleland2007_eco_map}
\alias{Cleland2007_eco_map}
\title{Ecological subsection maps, from Cleland et al. (2007).}
\format{
A list of three simple features objects (class "sf").
\describe{
  \item{province}{Province-level polygons}
  \item{sections}{Section-level polygons}
  \item{subsects}{Subsection-level polygons}
}
}
\usage{
Cleland2007_eco_map
}
\description{
Ecological subsection maps, from Cleland et al. (2007).
}
\references{
Cleland, D.T., Freeouf, J.A., Keys, J.E., Nowacki, G.J., Carpenter, C.A.,
    McNab, W.H., 2007. Ecological Subregions: Sections and Subsections for
    the conterminous United States.
    General Technical Report WO-76D 76D.
McNab, W.H., Cleland, D.T., Freeouf, J.A., Keys, J.E., Nowacki, G.J.,
    Carpenter, C.A., 2007. Description of ecological subregions:
    sections of the conterminous United States.
    General Technical Report WO-76B 76B, 1–82.
    https://doi.org/10.2737/WO-GTR-76B
}
\keyword{datasets}
