% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cleland_meta_df}
\alias{Cleland_meta_df}
\title{Cleland 2007 Province/Section/Subsection Names/Code Info}
\format{
A data.frame of 12 columns and 1233 rows, where each row corresponds
        to a single named subsection.
\describe{
  \item{domain}{Dry, Humid Temperate, Humid Tropical, or Water}
  \item{divsion}{Temperate Desert, Mediterranean, Prairie, i.e. biome}
  \item{province_name}{Short name of province}
  \item{province_code}{USFS Province Code, see `ECOSUBCD`` in FIA data}
  \item{province_description}{Long name of province}
  \item{section_id}{Section ID #, shortest name}
  \item{section_name}{Short name of section}
  \item{section_code}{USFS Section Code, see `ECOSUBCD`` in FIA data}
  \item{section_description}{Long name of section}
  \item{subsection_id}{Subsection ID #, shortest name}
  \item{subsection_name}{Short name of subsection}
  \item{subsection_code}{USFS Subection Code, see `ECOSUBCD`` in FIA data}
}
}
\usage{
Cleland_meta_df
}
\description{
Zone metadata pulled from 'Cleland2007_eco_map'.
}
\details{
See `Cleland_meta_df.Rmd` for details on generating this dataset. Not
updated when data format was changed to 'sf'.
}
\keyword{datasets}
