% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnipSingleCharacter.R
\name{SnipSingleCharacter}
\alias{SnipSingleCharacter}
\title{Trim a character vector.}
\usage{
SnipSingleCharacter(v, side = "front")
}
\description{
`SnipSingleCharacter` trims a character vector by 1 character for each vector element.
}
\examples{
SnipSingleCharacter(state.abb, side = 'back')
}
