% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{KeyClelandCode}
\alias{ScaleUpClelandCode}
\alias{NameClelandCode}
\alias{SnipSingleCharacter}
\title{Small utility functions for Cleland Ecoregions}
\usage{
KeyClelandCode(x, lvl = "province", long = F, rev = F)

ScaleUpClelandCode(x, in_lvl = "subsection", out_lvl = "section")

NameClelandCode(code)

SnipSingleCharacter(v, side = "front")
}
\description{
Some functions that are required for package processing or just useful to have
when working with the ClelandEcoregions dataset.
}
\section{Functions}{
\itemize{
\item \code{KeyClelandCode}: Key function for the FIA codes

\item \code{ScaleUpClelandCode}: Scales up the Cleland code

\item \code{NameClelandCode}: Prints the Cleland subsection name for a code

\item \code{SnipSingleCharacter}: Rrims a character by 1 character for each element
}}

